refactor(extend db upgrade): 删除掉一些之前合并版本的无关迁移文件。精简一下迁移文件，把改字段的统一都放到创建表里面去

+++ api/migrations_extend/versions/09633b4cf949_add_account_money_extend.py
@@ -22,8 +22,8 @@ def upgrade():
     op.create_table('account_money_extend',
                     sa.Column('id', models.types.StringUUID(), server_default=sa.text('uuid_generate_v4()'), nullable=False),
                     sa.Column('account_id', models.types.StringUUID(), nullable=False),
-                    sa.Column('total_quota', sa.Numeric(precision=10, scale=7), nullable=True),
-                    sa.Column('used_quota', sa.Numeric(precision=10, scale=7), nullable=True),
+                    sa.Column('total_quota', sa.Numeric(precision=16, scale=7), nullable=True),
+                    sa.Column('used_quota', sa.Numeric(precision=16, scale=7), nullable=True),
                     sa.Column('created_at', sa.DateTime(), server_default=sa.text('CURRENT_TIMESTAMP(0)'), nullable=False),
                     sa.Column('updated_at', sa.DateTime(), server_default=sa.text('CURRENT_TIMESTAMP(0)'), nullable=False),
                     sa.PrimaryKeyConstraint('id', name='account_money_pkey')
